requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.7
djangorestframework==3.16.0
et_xmlfile==2.0.0
idna==3.10
numpy==2.2.4
openpyxl==3.1.5
pandas==2.2.3
pillow==11.1.0
python-dateutil==2.9.0.post0
pytz==2025.2
razorpay==1.4.2
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.0
tzdata==2025.2
urllib3==2.4.0
gunicorn
